Add test suite, CI workflow, and fix all clippy warnings
- Add 20 unit tests for core logic (directive parsing, text wrapping, HTTP client)
- Add GitHub Actions workflow for tests on Linux/macOS/Windows
- Fix all clippy warnings to enable -D warnings in CI
- Add CLAUDE.md with contribution guidelines and architecture overview
- Add tempfile dev-dependency for file-based tests

Warnings fixed:
- Remove redundant imports
- Remove unused SubmissionResultMsg struct
- Remove unused get_ascii_art/display_ascii_art functions
- Fix int_plus_one, map_or, manual_strip, borrowed_box, redundant_closure
- Fix needless_return and unnecessary_cast

+++ src/views/result_page.rs
@@ -33,9 +33,7 @@ impl ResultPage {
 
         let num_lines = result_text.lines().count();
 
-        state.vertical_scroll_state = state
-            .vertical_scroll_state
-            .content_length(num_lines);
+        state.vertical_scroll_state = state.vertical_scroll_state.content_length(num_lines);
 
         state.horizontal_scroll_state = state.horizontal_scroll_state.content_length(max_width);
         state.animation_frame = 0;
@@ -70,7 +68,7 @@ impl ResultPage {
             .result_text
             .clone()
             .block(right_block)
-            .scroll((state.vertical_scroll as u16, state.horizontal_scroll as u16));
+            .scroll((state.vertical_scroll, state.horizontal_scroll));
         result_text.render(right, buf);
     }
